refactor: extract useTaskReorder hook to deduplicate task reorder logic
