+++ src/tasks/components/TasksSection/TasksSection.tsx
@@ -4,9 +4,9 @@ import { Button } from "src/common/components/Button/Button";
 import { useCurrentPocketbookId } from "src/pocketbooks/hooks/useCurrentPocketbookId";
 import { TaskEditor } from "src/tasks/components/TaskEditor/TaskEditor";
 import { useCreateTask } from "src/tasks/hooks/useCreateTask";
-import { useUpdateTask } from "src/tasks/hooks/useUpdateTask";
+import { useTaskReorder } from "src/tasks/hooks/useTaskReorder";
 import type { Colour } from "src/colours/Colour.type";
-import type { Task, TasksGroup } from "src/tasks/Task.type";
+import type { TasksGroup } from "src/tasks/Task.type";
 
 type TasksSectionProps = {
   taskGroup: TasksGroup;
@@ -23,7 +23,7 @@ export const TasksSection = ({
   const [isTitleHovered, setIsTitleHovered] = useState(false);
   const [newTaskFocusId, setNewTaskFocusId] = useState<string | null>(null);
   const { createTask } = useCreateTask();
-  const { updateTask } = useUpdateTask();
+  const { getMoveCallbacks } = useTaskReorder();
   const { pocketbookId } = useCurrentPocketbookId();
   const handledToolbarTriggerRef = useRef(0);
 
@@ -47,29 +47,6 @@ export const TasksSection = ({
       setNewTaskFocusId(createdTask.id);
     }
   }, [createTask, note]);
-
-  const swapTaskOrder = useCallback(
-    (taskA: Task, taskB: Task) => {
-      updateTask({ taskId: taskA.id, updateTaskData: { ...taskA, sortOrder: taskB.sortOrder } });
-      updateTask({ taskId: taskB.id, updateTaskData: { ...taskB, sortOrder: taskA.sortOrder } });
-    },
-    [updateTask],
-  );
-
-  const getMoveCallbacks = useCallback(
-    (index: number, tasks: Task[]) => {
-      const onMoveUp =
-        index > 0
-          ? () => swapTaskOrder(tasks[index], tasks[index - 1])
-          : undefined;
-      const onMoveDown =
-        index < tasks.length - 1
-          ? () => swapTaskOrder(tasks[index], tasks[index + 1])
-          : undefined;
-      return { onMoveUp, onMoveDown };
-    },
-    [swapTaskOrder],
-  );
 
   // Create a new no-note task whenever the toolbar plus button fires.
   useEffect(() => {
